Crate zkp_client

zkp_client includes modules and functions used to communicate with the zkp_server These include functionalities for:

Generating the commits r1, r2 using secret x (set via env variables)

Generating random k value and computing y1 and y2 by g^K and h^K

zkp flow

create_commits

register_user_with_server

prove authenticity using s = k - c.x where c is the challenge from server

continue to communicate with server once logged in

Constraints

Following Constraints are set for convenience during exponential operations Example:

pub const MAX_USERNAME_LEN: usize = 50;
// min 'k' value set so that "s = k - c.x" is always positive.
pub const MIN_SEED_VAL: u32 = 100;
// max 'k' value. used during g^K and h^K operations
pub const MAX_SEED_VAL: u32 = 125;
// used when computing y1 = g^x and y2 = h^x
pub const MAX_SECRET_VAL: u32 = 25;

Modules

auth

seed

user

Constants

MAX_SECRET_VAL

MAX_SEED_VAL

MAX_USERNAME_LEN

MIN_SEED_VAL

Functions

connect

Uses TCPStream::connect to establish connection to server and returns a TcpStream instance if successful.

init_logger

Env Logger level info. Set $RUST_LOG env variable accordingly to change severity levels

init_zkp_flow

create_commits

register_user_with_server

prove authenticity using s = k - c.x where c is the challenge from server

continue to communicate with server once logged in

login_callback

print_errors

Prints any errors that occur during env variables validation

Type Definitions

Callback